Delete unnecessary files
