--- apps/vs-code-designer/src/app/commands/createProject/createProjectSteps/projectCreateStepBase.ts
@@ -1,49 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-import { localize } from '../../../../localize';
-import { gitInit, isGitInstalled, isInsideRepo } from '../../../utils/git';
-import { AzureWizardExecuteStep, callWithTelemetryAndErrorHandling } from '@microsoft/vscode-azext-utils';
-import type { IActionContext } from '@microsoft/vscode-azext-utils';
-import type { IProjectWizardContext } from '@microsoft/vscode-extension-logic-apps';
-import * as fse from 'fs-extra';
-import type { Progress } from 'vscode';
-import { workflowCodeTypeForTelemetry } from '../../../utils/codeful/utils';
-
-export abstract class ProjectCreateStepBase extends AzureWizardExecuteStep<IProjectWizardContext> {
-  public priority = 10;
-
-  public abstract executeCore(
-    context: IProjectWizardContext,
-    progress: Progress<{ message?: string | undefined; increment?: number | undefined }>
-  ): Promise<void>;
-
-  public shouldExecute(_context: IProjectWizardContext): boolean {
-    return true;
-  }
-
-  public async execute(
-    context: IProjectWizardContext,
-    progress: Progress<{ message?: string | undefined; increment?: number | undefined }>
-  ): Promise<void> {
-    context.telemetry.properties.projectLanguage = context.language;
-    context.telemetry.properties.projectRuntime = context.version;
-    context.telemetry.properties.openBehavior = context.openBehavior;
-
-    progress.report({ message: localize('creating', 'Creating new project...') });
-    await fse.ensureDir(context.projectPath);
-
-    await this.executeCore(context, progress);
-
-    if ((await isGitInstalled(context.workspacePath)) && !(await isInsideRepo(context.workspacePath))) {
-      await gitInit(context.workspacePath);
-    }
-
-    // OpenFolderStep sometimes restarts the extension host. Adding a second event here to see if we're losing any telemetry
-    await callWithTelemetryAndErrorHandling('azureLogicAppsStandard.createProjectStarted', (startedContext: IActionContext) => {
-      context.telemetry.properties.workflowCodeType = workflowCodeTypeForTelemetry(context.isCodeless);
-      Object.assign(startedContext, context);
-    });
-  }
-}
