--- apps/vs-code-designer/src/app/commands/createWorkspace/createWorkspaceSteps/openFolderStep.ts
@@ -1,61 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-import { getContainingWorkspace } from '../../../utils/workspace';
-import { AzureWizardExecuteStep } from '@microsoft/vscode-azext-utils';
-import type { IProjectWizardContext } from '@microsoft/vscode-extension-logic-apps';
-import { OpenBehavior } from '@microsoft/vscode-extension-logic-apps';
-import * as fs from 'fs';
-import { commands, Uri, workspace } from 'vscode';
-import { extensionCommand } from '../../../../constants';
-
-export class OpenFolderStep extends AzureWizardExecuteStep<IProjectWizardContext> {
-  public priority = 250;
-
-  /**
-   * Determines whether this step should be executed based on the user's input.
-   * @param context The context object for the project wizard.
-   * @returns A boolean value indicating whether this step should be executed.
-   */
-  public shouldExecute(context: IProjectWizardContext): boolean {
-    return !!context.openBehavior && context.openBehavior !== OpenBehavior.alreadyOpen && context.openBehavior !== OpenBehavior.dontOpen;
-  }
-
-  /**
-   * Executes the step to open the folder in Visual Studio Code.
-   * @param context The context object for the project wizard.
-   * @returns A Promise that resolves to void.
-   */
-  public async execute(context: IProjectWizardContext): Promise<void> {
-    const openFolders = workspace.workspaceFolders || [];
-    let workspaceUri: Uri;
-
-    // Check if .code-workspace file exists in project path
-    const workspaceFilePath = context.workspaceFilePath;
-    context.workspaceFolder = getContainingWorkspace(workspaceFilePath);
-    if (fs.existsSync(workspaceFilePath)) {
-      workspaceUri = Uri.file(workspaceFilePath);
-    } else {
-      workspaceUri = Uri.file(context.workspacePath);
-    }
-
-    // Check if user has selected to add folder to workspace and update workspace accordingly
-    if (context.openBehavior === OpenBehavior.addToWorkspace && openFolders.length === 0) {
-      context.openBehavior = OpenBehavior.openInCurrentWindow;
-    }
-
-    if (context.openBehavior === OpenBehavior.addToWorkspace) {
-      if (!workspaceUri.path.endsWith('.code-workspace')) {
-        workspace.updateWorkspaceFolders(openFolders.length, 0, { uri: workspaceUri });
-      }
-    } else {
-      // Open folder using executeCommand method of commands object with vscode.openFolder command
-      await commands.executeCommand(
-        extensionCommand.vscodeOpenFolder,
-        workspaceUri,
-        context.openBehavior === OpenBehavior.openInNewWindow /* forceNewWindow */
-      );
-    }
-  }
-}
